moves tuple functionality into thrust/detail/tuple_detail, includes tuple_io.h and tuple_cat.h from tuple.h, fixes syntax errors and warnings in tuple_cat.h/tuple_helpers.h
